+++ mazikeen/GeneratorUtils.py
@@ -1,15 +1,4 @@
-from yaml.loader import SafeLoader
-
 from mazikeen.GeneratorException import GeneratorException
-
-
-class SafeLineLoader(SafeLoader):
-    def construct_mapping(self, node, deep=False):
-        mapping = super(SafeLineLoader, self).construct_mapping(node, deep=deep)
-        # Add 1 so line numbering starts at 1
-        mapping['__line__'] = node.start_mark.line + 1
-        return mapping
-        
 
 def getYamlInt(data, line, field):
     if (not isinstance(data, int)):